V22

+++ src/user/Kunde.java
@@ -5,7 +5,7 @@ import src.rtd.*;
 
 public class Kunde extends User {
 
-	public int place(int value) {
+	public double place(int value) {
 		// if address 0 then RTDR
 
 		// or
